pyproject.toml
[tool.poetry]
name = "deepseek-daily-newsletter"
version = "0.1.0"
description = ""
authors = ["Mohit Bhole <mohitb.steam@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10"
PyYAML = "*"
requests = "*"
schedule = "*"
python-dotenv = "*"
openai = "^2.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-watch = "^4.2.0"

[tool.poetry.group.dev]
optional = true

[project]
name = "deepseek-daily-newsletter"
version = "0.1.0"
description = ""
authors = [{name = "Mohit Bhole",email = "mohitb.steam@gmail.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["PyYAML", "requests", "schedule", "pytest", "python-dotenv", "openai"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
